backend/pyproject.toml
[project]
name = "mine-chan-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.13.3",
    "annotated-types==0.6.0",
    "anyio==4.3.0",
    "asyncpg>=0.29.0",
    "cachecontrol==0.14.0",
    "cachetools==5.3.3",
    "certifi==2024.2.2",
    "cffi==1.16.0",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "colorama==0.4.6",
    "cryptography==42.0.5",
    "fastapi==0.110.0",
    "firebase-admin==6.5.0",
    "google-api-core==2.18.0",
    "google-api-python-client==2.123.0",
    "google-auth-httplib2==0.2.0",
    "google-auth==2.29.0",
    "google-cloud-core==2.4.1",
    "google-cloud-firestore==2.15.0",
    "google-cloud-storage==2.16.0",
    "google-crc32c==1.5.0",
    "google-resumable-media==2.7.0",
    "googleapis-common-protos==1.63.0",
    "grpcio-status==1.62.1",
    "grpcio==1.62.1",
    "h11==0.14.0",
    "httplib2==0.22.0",
    "httptools==0.6.1",
    "idna==3.6",
    "iniconfig==2.0.0",
    "msgpack==1.0.8",
    "numpy==1.26.4",
    "packaging==24.0",
    "pluggy==1.4.0",
    "proto-plus==1.23.0",
    "protobuf==4.25.3",
    "psycopg2-binary>=2.9.9",
    "pyasn1-modules==0.3.0",
    "pyasn1==0.5.1",
    "pycparser==2.21",
    "pydantic-core==2.16.3",
    "pydantic==2.6.4",
    "pyjwt==2.8.0",
    "pyparsing==3.1.2",
    "pytest==8.1.1",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.1",
    "redis==5.0.3",
    "requests==2.31.0",
    "rsa==4.9",
    "sniffio==1.3.1",
    "sqlalchemy>=2.0.35",
    "starlette==0.36.3",
    "typing-extensions==4.10.0",
    "uritemplate==4.1.1",
    "urllib3==2.2.1",
    "uvicorn==0.29.0",
    "watchfiles==0.21.0",
    "websockets==12.0",
]
